### 2024-12-04 0f7f42d
Add guidelines for contributing
