### 2024-12-04 #7462 @kmlefran
@Kevin-Mattheus-Moerman Thanks for informing me. I have added CONTRIBUTING.md and linked to it in the README
