### 2025-02-10 #7462 @HaoZeke
Hi @BernardoFerreira 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
From your work on CRATE it would be good to have your opinion.
