### 2025-02-10 #7462 @HaoZeke
Actually as an AiiDA plugin it'd also be fantastic to have anyone from @aiidateam take a look if possible.
