### 2025-05-20 #7462 @yw-fang
> Hi [@yw-fang](https://github.com/yw-fang) 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

Thank you for reaching out to me. However, I have multiple papers from other journals to review in the coming weeks. Given this, I would recommend other scientists. Bonan @zhubonan is an expert on AiiDA plugins who comes to mind right now.
