### 2025-05-21 #7462 @mbagheri20
> Hi @mbagheri20 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

Hi @HaoZeke, Thank you for reaching out. I'll be busy until mid June and I can start reviewing after that.
