### 2025-05-22 #7462 @HaoZeke
Hi @zhubonan 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

@yw-fang thanks for the recommendation :)
@mbagheri20 thanks for the prompt response, I'll keep looking but if it gets to mid June it'd be great to have your inputs :)
