### 2025-05-28 #8296 @editorialbot
[REVIEW]: aiida-aimall: A Python package for automating workflows for AIMAll software
**Submitting author:** @kmlefran (<a href="http://orcid.org/0000-0001-5291-1795">Kevin Lefrancois-Gagnon</a>)
**Repository:** https://github.com/kmlefran/aiida-aimall
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.0.5
