+++ paper.bib
@@ -26,6 +26,7 @@ title = {
   doi = {10.1021/acs.jcim.3c00987}
 }
 @software{AIMAll,
+  title        ={AIMAll},
   author       = {Todd A. Keith},
   year         = 2019,
   publisher    = {TK Gristmill Software},
