### 2025-07-04 b32b607
Restrict aiida-submission-controller to not use the latest versions.
Will update with code to account for the latest soon
