### 2025-07-13 #8296 @zhubonan
I can successfully install the package and run the tests. The code looks good to me as well, just two suggestions:
1. `group_decomposition` needs to be added to the test dependencies needs in the pyproject.toml.
2. There should be clearer instructions on how to contribute and report issues in the documentation (e.g. code guidelines and raise PR/Issues on GitHub for example).
