### 2025-07-14 65ebfde
Add group_decomposition to testing dependencies
