### 2025-07-14 bc81b76
Improve developer guideline documentation
