+++ docs/source/developer_guide/index.md
@@ -1,0 +1,53 @@
+# Contributing
+
+(contributing)=
+
+First off, thanks for taking the time to contribute!
+
+All types of contributions are encouraged and valued.  Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions.
+
+```{toctree}
+:hidden: true
+:maxdepth: 1
+
+bug_reports.md
+developer_guide.md
+
+```
+
+::::{grid} 2
+:gutter: 3
+
+:::{grid-item-card} {fa}`rocket;mr-1` Bug Reports
+:text-align: center
+:shadow: md
+
++++
+
+```{button-ref} bug_reports
+:ref-type: doc
+:click-parent:
+:expand:
+:color: primary
+:outline:
+```
+
+:::
+
+:::{grid-item-card} {fa}`rocket;mr-1` Developer Guide
+:text-align: center
+:shadow: md
+
++++
+
+```{button-ref} developer_guide
+:ref-type: doc
+:click-parent:
+:expand:
+:color: primary
+:outline:
+```
+
+:::
+
+::::
