### 2025-07-14 3842ab3
Update sigstore action and remove paper from main
