+++ paper.bib
@@ -1,72 +1,0 @@
-@article{aiida,
-  title        = {
-    {AiiDA 1.0, a scalable computational infrastructure for automated
-    reproducible workflows and data provenance }
-  },
-  author       = {
-    Leopold Talirz and Leonid Kahle and Rico Häuselmann and Dominik Gresch and
-    Tiziano Müller and Aliaksandr V. Yakutovich and Casper W. Andersen and
-    Francisco F. Ramirez and Carl S. Adorf and Fernando Gargiulo and Snehal
-    Kumbhar and Elsa Passaro and Conrad Johnston and Andrius Merkys and Andrea
-    Cepellotti and Nicolas Mounet and Nicola Marzari and Boris Kozinsky and
-    Giovanni Pizzi
-  },
-  year         = 2020,
-  journal      = {Scientific Data},
-  volume       = 7,
-  doi          = {10.1038/s41597-020-00638-4}
-}
-@article{klg,
-title = {
-  {Toward Universal Substituent Constants: Relating QTAIM Functional Group Descriptors to Substituent Effect Proxies}},
-  author = { Kevin M. Lefrancois-Gagnon and Robert C. Mawhinney},
-  year          = 2023,
-  journal = {Journal of Chemical Information and Modeling},
-  volume = 63,
-  doi = {10.1021/acs.jcim.3c00987}
-}
-@software{AIMAll,
-  title        ={AIMAll},
-  author       = {Todd A. Keith},
-  year         = 2019,
-  publisher    = {TK Gristmill Software},
-  url          = {aim.tkgristmill.com},
-  version      = {19.10.12}
-}
-@misc{gaussian,
-  title        = {Gaussian˜16 {R}evision {C}.01},
-  author       = {
-    M. J. Frisch and G. W. Trucks and H. B. Schlegel and G. E. Scuseria and M.
-    A. Robb and J. R. Cheeseman and G. Scalmani and V. Barone and G. A.
-    Petersson and H. Nakatsuji and X. Li and M. Caricato and A. V. Marenich and
-    J. Bloino and B. G. Janesko and R. Gomperts and B. Mennucci and H. P.
-    Hratchian and J. V. Ortiz and A. F. Izmaylov and J. L. Sonnenberg and D.
-    Williams-Young and F. Ding and F. Lipparini and F. Egidi and J. Goings and
-    B. Peng and A. Petrone and T. Henderson and D. Ranasinghe and V. G.
-    Zakrzewski and J. Gao and N. Rega and G. Zheng and W. Liang and M. Hada and
-    M. Ehara and K. Toyota and R. Fukuda and J. Hasegawa and M. Ishida and T.
-    Nakajima and Y. Honda and O. Kitao and H. Nakai and T. Vreven and K.
-    Throssell and Montgomery, {Jr.}, J. A. and J. E. Peralta and F. Ogliaro and
-    M. J. Bearpark and J. J. Heyd and E. N. Brothers and K. N. Kudin and V. N.
-    Staroverov and T. A. Keith and R. Kobayashi and J. Normand and K.
-    Raghavachari and A. P. Rendell and J. C. Burant and S. S. Iyengar and J.
-    Tomasi and M. Cossi and J. M. Millam and M. Klene and C. Adamo and R. Cammi
-    and J. W. Ochterski and R. L. Martin and K. Morokuma and O. Farkas and J.
-    B. Foresman and D. J. Fox
-  },
-  year         = 2016,
-  note         = {Gaussian Inc. Wallingford CT}
-}
-@software{aiidagaussian,
-  title        = {nanotech-empa/aiida-gaussian: v2.1.0},
-  author       = {
[…]
