+++ paper.md
@@ -1,105 +1,0 @@
----
-title: 'aiida-aimall: A Python package for automating workflows for AIMAll software'
-tags:
-  - Python
-  - AiiDA
-  - Quantum Theory of Atoms in Molecules
-  - AIMAll
-  - Computational chemistry
-authors:
-  - name: Kevin M. Lefrancois-Gagnon
-    orcid: 0000-0001-5291-1795
-    equal-contrib: true
-    affiliation: 1 # (Multiple affiliations must be quoted)
-  - name: Robert C. Mawhinney
-    orcid: 0000-0002-6077-2403
-    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
-    affiliation: 1
-affiliations:
- - name: Lakehead University, Thunder Bay, ON, Canada
-   index: 1
-date: 14 July 2024
-bibliography: paper.bib
-
----
-
-# Summary
-
-Since its introduction by Richard Bader, the Quantum Theory of Atoms in Molecules
-(QTAIM) has become a useful tool for computational chemists. This Python package
-provides plugins for a common QTAIM software, AIMAll, for the AiiDA Python
-infrastructure. `aiida-aimall` is an essential tool for ensuring reproducible
-calculations, with full generation history. Workflows are also provided to interface
-`AIMAll` software with any quantum chemistry package that can be run through the command line,
-so long as it generates the input files required by `AIMAll`.
-
-# Statement of need
-
-`aiida-aimall` is a Python package based on the AiiDA [@aiida] infrastructure designed
-to assist users with generating inputs for AIMAll software [@AIMAll]. The goal of
-the AiiDA infrastructure is, in part, to ensure data provenance and calculation
-reproducibility. While `aiida-aimall` has been developed primarily for interface
-with Gaussian software outputs [@gaussian], a versatile workflow enabling interface with
-other quantum chemistry packages is also made available. To the best of the authors' knowledge, no tool for automatically linking the output of quantum chemistry packages with AIMAll while tracking data provenance exists. Before this tool, individuals would manually provide output files from one software to the other. This tool significantly simplifies complex multi-step workflows such as substituent parameter generation into one simple step, easing the burden on the end user and allowing for high-throughput calculations.
-
-Through a variety of workflows that can start with an .xyz file, AiiDA `StructureData`, or even with
-a SMILES string of a molecule, `aiida-aimall` provides a variety of use cases for automating complex workflows. Additionally, tools to ensure that computers are not overloaded through
-too many simultaneous processes are made availabe through classes of `FromGroupSubmissionController`s
-from `aiida-submission-controller` to limit active processes.
-
-# Features
-`aiida-aimall` contains many different classes from `aiida` tailored to ensure ease of use of
-AIMAll calculations. Numerous features provided by `aiida-aimall` are described in full on the [documentation webpage hosted on ReadTheDocs](https://aiida-aimall.readthedocs.io/en/latest/). A brief description of main features is provided here.
-
-## Running Simple AIMAll Calculations
-
-The simplest functionality provided by `aiida-aimall` is running AIMAll calculations. All AIMAll calculations utilize the `AimqbParameters` datatype provided by `aiida-aimall`. The `AimqbParameters` datatype
-is a validator for `AIMAll` command line input. Command line parameters are to be provided as a dictionary,
-then `AimqbParameters` ensures that the parameters match options available for AIMAll software as
-[defined on the software website](https://aim.tkgristmill.com/manual/aimqb/aimqb.html), and that the
-correct data type is provided for each parameter. In this way, `AimqbParameters` verifies the provided input
-to AIMAll calculations prior to launch of the calculation. These parameters, along with `SinglefileData` of a valid AIMAll input file, a `Code` object for AIMAll software, and relevant metadata are provided to an `AimqbCalculation`.
-
[…]
