### 2025-07-14 #8296 @kmlefran
@zhubonan Thanks for the feedback. I have updated the testing dependencies and revamped the developer guide in the documentation to address issues and pull requests. The updated documentation is live on ReadTheDocs and I've published package version 1.0.3 on PyPI with updated dependencies
