### 2025-10-20 #8296 @rkurchin
(pinged @HaoZeke on Slack)
