### 2025-10-31 9b0ecf3
Add image and address reviewer comment
