### 2025-10-31 1a98731
Fix typo and figure reference
