### 2025-11-15 #8296 @HaoZeke
@kmlefran yup I think we can move forward with this, can you provide the DOI after the Zenodo archive?
