+++ src/aiida_aimall/__init__.py
@@ -1,4 +1,4 @@
 """Integration of AIMQB and aiida
 """
 
-__version__ = "1.0.4"
+__version__ = "1.0.5"
