+++ .github/workflows/draft-pdf.yml
@@ -13,6 +13,8 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+        with:
+          ref: ${{ github.head_ref || github.ref_name }}
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
