### 2025-11-16 997ac0b
Draft PDF fix
