### 2025-11-16 #8296 @kmlefran
I've created a new release, version 1.0.5 that includes the paper and the zenodo doi is
10.5281/zenodo.17620473
