### 2025-11-28 #8296 @kmlefran
@HaoZeke what are the next steps? Is the zenodo release set up correctly?
