### 2025-12-19 #8296 @rkurchin
Okay, I ran this error through ChatGPT and actually got a suggestion for the possible issue; it seems it's likely related to the table formatting, and it suggested two possible fixes to replace the table:

### option 1

```
+---------------------------------+-----------------+------------------------------+
| Workflow                        | Entry Point     | Purpose                      |
+---------------------------------+-----------------+------------------------------+
| `QMToAIMWorkchain`              | aimall.qmtoaim  | Run a general computational  |
|                                 |                 | chemistry software and link  |
|                                 |                 | it to an AIMAll              |
|                                 |                 | calculation                  |
+---------------------------------+-----------------+------------------------------+
| `GenerateWFXToAIMWorkchain`     | aimall.wfxtoaim | Take non-standard AIMAll     |
|                                 |                 | input files, and run AIMAll  |
+---------------------------------+-----------------+------------------------------+
| `GaussianToAIMWorkChain`        | aimall.g16toaim | Run a Gaussian calculation   |
|                                 |                 | and automatically run        |
|                                 |                 | an AIMAll calculation on     |
|                                 |                 | its outputs                  |
+---------------------------------+-----------------+------------------------------+
| `SubstituentParameterWorkChain` | aimall.subparam | Compute substituent          |
|                                 |                 | properties defined by        |
|                                 |                 | the authors automatically   |
+---------------------------------+-----------------+------------------------------+
```

### option 2
```
| Workflow | Entry Point | Purpose |
|---------|-------------|---------|
| `QMToAIMWorkchain` | aimall.qmtoaim | Run a general computational chemistry software and link it to an AIMAll calculation |
| `GenerateWFXToAIMWorkchain` | aimall.wfxtoaim | Take non-standard AIMAll input files, and run AIMAll |
| `GaussianToAIMWorkChain` | aimall.g16toaim | Run a Gaussian calculation and automatically run an AIMAll calculation on its outputs |
| `SubstituentParameterWorkChain` | aimall.subparam | Compute substituent properties defined by the authors automatically |
```

@kmlefran, can you try this out and confirm the content and formatting look okay and then ping and we can try processing it again?
