### 2025-12-19 #8296 @kmlefran
I’ll try it out later tonight. @rkurchin, do I need to publish a new version with the change and updated Zenodo archive to test the GitHub action, or is just updating main sufficient?
