### 2026-01-17 #8296 @HaoZeke
@rkurchin I thought this was already ready for acceptance.. Odd.
