### 2026-01-17 #8296 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
