### 2026-01-24 #8296 @kmlefran
@Kevin-Mattheus-Moerman I have altered the Zenodo archive title at https://zenodo.org/records/17620473 and fixed the typos in paper.md on the paper branch in the repository.

Please let me know if other action is required
